+++ spec/models/form_submission_spec.rb
@@ -4,6 +4,8 @@ describe FormSubmission do
   let(:model) { described_class }
   let(:klass) { model.to_s.underscore.to_sym }
 
+  it_behaves_like :emailable
+
   describe 'validations' do
     describe 'presence' do
       it(:form) { expect(build(klass, form: nil)).not_to be_valid }